Reformat mock.ts
