+++ src/app/problems/_tests/problem-create/java-create-snippet.component.spec.ts
@@ -12,7 +12,7 @@ import {CourseServiceMock} from "@test/course.service.mock";
 import {QuestionService} from "@app/problems/_services/question.service";
 import {QuestionServiceMock} from "@app/problems/_tests/question.service.mock";
 import {ReactiveFormsModule} from "@angular/forms";
-import {MOCK_COURSE} from "@app/problems/_tests/mock";
+import {MOCK_COURSES} from "@test/mock";
 
 describe('JavaCreateSnippetComponent', () => {
     let component: JavaCreateSnippetComponent;
@@ -42,7 +42,7 @@ describe('JavaCreateSnippetComponent', () => {
 
     it('courseSelectedById', () => {
         component.courseSelectedById(0);
-        expect(component.events).toEqual(MOCK_COURSE.events);
+        expect(component.events).toEqual(MOCK_COURSES[0].events);
     });
 
     it('java create question', () => {
