+++ src/app/problems/_tests/problem-create/mcq-create-snippet.component.spec.ts
@@ -6,13 +6,13 @@ import {CategoryService} from "@app/_services/api/category.service";
 import {CategoryServiceMock} from "@test/category.service.mock";
 import {CourseService} from "@app/_services/api/course/course.service";
 import {CourseServiceMock} from "@test/course.service.mock";
-import {MOCK_CATEGORIES, MOCK_COURSE} from "@app/problems/_tests/mock";
 import {CKEditorModule} from "@ckeditor/ckeditor5-angular";
 import {CkEditorComponent} from "@app/problems/ck-editor/ck-editor.component";
 import {JsonEditorComponent} from "@app/problems/json-editor/json-editor.component";
 import {ReactiveFormsModule} from "@angular/forms";
 import {QuestionService} from "@app/problems/_services/question.service";
 import {QuestionServiceMock} from "@app/problems/_tests/question.service.mock";
+import {MOCK_CATEGORIES, MOCK_COURSES} from "@test/mock";
 
 describe('McqCreateSnippetComponent', () => {
     let component: McqCreateSnippetComponent;
@@ -41,7 +41,7 @@ describe('McqCreateSnippetComponent', () => {
     });
 
     it('courses and categories set', () => {
-        expect(component.courses).toEqual([MOCK_COURSE]);
+        expect(component.courses).toEqual(MOCK_COURSES);
         expect(component.categories).toEqual(MOCK_CATEGORIES);
     });
 
@@ -80,6 +80,6 @@ describe('McqCreateSnippetComponent', () => {
 
     it('courseSelectedById', () => {
         component.courseSelectedById(0);
-        expect(component.events).toEqual(MOCK_COURSE.events);
+        expect(component.events).toEqual(MOCK_COURSES[0].events);
     });
 });
